requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0